fix(core): keep a tracked row's card when its first cell is typed in
A row's `w:trPr/w:ins` and `w:cellIns` markers live outside any paragraph and
anchor to the row's first one, so the local review patch counted them as
paragraph-local, dropped them, and the paragraph re-walk had nothing to put
back. The row stayed painted as a proposal with no decision behind it.

Extracts the patch rules to their own module: tree-session.ts was already over
its line cap.
