+++ .changeset/tracked-row-card-survives-typing.md
@@ -1,0 +1,5 @@
+---
+'@docx-editor.dev/core': patch
+---
+
+Typing in a tracked table row no longer drops that row's tracked-change card, so the row insertion stays acceptable and rejectable.
